quotamanager/requirements.txt
PyYAML
shade
